+++ addon/query/indexeddb-adapter.js
@@ -3,7 +3,6 @@
 */
 
 import RSVP from 'rsvp';
-import EmberMap from '@ember/map';
 import { getOwner } from '@ember/application';
 import { warn } from '@ember/debug';
 import FilterOperator from './filter-operator';
@@ -12,7 +11,7 @@ import BaseAdapter from './base-adapter';
 import JSAdapter from 'ember-flexberry-data/query/js-adapter';
 import Information from '../utils/information';
 import getSerializedDateValue from '../utils/get-serialized-date-value';
-import Dexie from 'npm:dexie';
+import Dexie from 'dexie';
 import Queue from '../utils/queue';
 
 /**
@@ -76,7 +75,7 @@ export default class extends BaseAdapter {
       };
 
       let getDetailsHashMap = function(data, primaryKeyName) {
-        let ret = EmberMap.create();
+        let ret = new Map();
         let dataLength = data.length;
         for (let i = 0; i < dataLength; i++) {
           let obj = data[i];