+++ app/browserify.js
@@ -1,8 +1,0 @@
-/* eslint-disable no-unused-vars */
-
-// Requirement of ember-browserify.
-// In order to use NPM pacakges inside the addon, we have to import
-// them from somewhere in /app directory.
-// See: https://github.com/ef4/ember-browserify#using-ember-browserify-in-addons
-import Dexie from 'npm:dexie';
-import uuid from 'npm:node-uuid';